backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
celery>=5.3.0
redis>=5.0.0
python-telegram-bot>=20.0
sentry-sdk>=1.39.0
python-dotenv>=1.0.0
Pillow>=10.0.0
django-storages>=1.14.0
boto3>=1.34.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
prometheus-client==0.19.0
alembic==1.12.1
pydantic==2.5.2
pydantic-settings==2.1.0
django-redis>=5.4.0
django-filter>=23.5
drf-yasg>=1.21.7
django-allauth>=0.57.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
gunicorn>=21.2.0
whitenoise>=6.6.0
requests>=2.31.0
PyJWT>=2.8.0
django-environ>=0.11.2
google-api-python-client>=2.113.0
